Restructure file system
