+++ src/api/UserAccount/deleteUser.js
