+++ src/api/game/room.class.js
